services/authentication_service/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
Jinja2==2.10.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
flask_restx==0.2.0
protobuf==3.20.0
grpcio==1.44.0
grpcio-tools==1.44.0
